add new endpoint, manager, and dao for paged listing activity metdata
