+++ chpl/chpl-resources/src/main/resources/errors.properties
@@ -20,6 +20,10 @@ access.pendingSurveillances=Pending surveillances may only be viewed by the ONC-
 #activity errors
 activity.missingStartHasEnd=An end date has been provided which means a start date is also required.
 activity.missingEndHasStart=A start date has been provided which means an end date is also required.
+activity.startDateAfterEndDate=Cannot search for activity with the start date after the end date.
+activity.negativePageNum=The page number must be 0 or greater.
+activity.negativePageSize=The page size must be greater than 0.
+activity.pageSizeOverMax=The page size cannot be greater than %s.
 
 #announcement errors
 announcement.notFound=Announcement not found.
